CLI is out the window
I wanted to do a nice text-based interface with checkboxes and scrollable lists that the user could move a cursor up/down to select from (transactions/accounts/etc), but the library I was using won't play nice (or at all) with IntelliJ IDEA's console output so it's going to be too much hassle. I might come back to it later if there's time.