Convert HashMap of input details to TreeMap (sorted alphabetically by field name)
Doesn't affect the functional outcome of the SQL update statement, but using a HashMap meant the Key/Value pairs were being added to the statement String in random order each time which made it impossible to write automated tests verifying the output statement against the expected result. This resolves that issue.